\name{RIFS-package}
\alias{RIFS-package}
\alias{RIFS}
\docType{package}
\title{Random Iterated Function System (RIFS)}
\description{
Pointwise generation and display of attractors (prefractals) of the random iterated function system (RIFS) for various combinations of probabilistic and geometric parameters of some fixed point sets (protofractals), described by Bukhovets A.G. (2012) <doi:10.1134/S0005117912020154>.
}
\details{
\tabular{ll}{
Package: \tab RIFS\cr
Type: \tab Package\cr
Version: \tab 0.1.6\cr
Date: \tab 2022-05-08\cr
License: \tab GPL-3\cr
}
\code{plotR2pre()} function draws a prefractal set in R^2.\cr
\code{preRIFS()} function generates a sample of fractal points (a prefractal points) in R^n with a random iterated function system (RIFS).\cr
\code{preRSum0()} function generates a sample of fractal points (a prefractal points) in R^n with a matrix of random sums of a numerical series.\cr
\code{R2ngon()} function generates a regular polygonal set in R^2.\cr
}
\author{Pavel V. Moskalev, Alexey G. Bukhovets and Tatyana Ya. Biruchinskay}
