\name{RIFS-package}
\alias{RIFS-package}
\alias{RIFS}
\docType{package}
\title{Random Iterated Function System (RIFS)}
\description{
RIFS package provides functionality for generating & plotting prefractal sets in R^n with various protofractal sets, probability distribution, and partition coefficient for iterative segments.
}
\details{
\tabular{ll}{
Package: \tab RIFS\cr
Type: \tab Package\cr
Version: \tab 0.1-5\cr
Date: \tab 2012-06-04\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
\code{plotR2pre()} function draws a prefractal set in R^2.\cr
\code{preRIFS()} function generates a sample of fractal points (a prefractal points) in R^n with a random iterated function system (RIFS).\cr
\code{preRSum0()} function generates a sample of fractal points (a prefractal points) in R^n with a matrix of random sums of a numerical series.\cr
\code{R2ngon()} function generates a regular polygonal set in R^2.\cr
}
\author{Pavel V. Moskalev, Alexey G. Bukhovets and Tatyana Ya. Biruchinskay}
